+++ plugins/plotly-express/src/js/src/PlotlyExpressChart.tsx
@@ -1,0 +1,46 @@
+import React, { useEffect, useRef, useState } from 'react';
+import Plotly from 'plotly.js-dist-min';
+import { Chart } from '@deephaven/chart';
+import { type WidgetComponentProps } from '@deephaven/plugin';
+import { useApi } from '@deephaven/jsapi-bootstrap';
+import PlotlyExpressChartModel from './PlotlyExpressChartModel.js';
+import { useHandleSceneTicks } from './useHandleSceneTicks.js';
+
+export function PlotlyExpressChart(
+  props: WidgetComponentProps
+): JSX.Element | null {
+  const dh = useApi();
+  const { fetch } = props;
+  const containerRef = useRef<HTMLDivElement>(null);
+  const [model, setModel] = useState<PlotlyExpressChartModel>();
+
+  useEffect(() => {
+    let cancelled = false;
+    async function init() {
+      const widgetData = await fetch();
+      if (!cancelled) {
+        setModel(new PlotlyExpressChartModel(dh, widgetData, fetch));
+      }
+    }
+
+    init();
+
+    return () => {
+      cancelled = true;
+    };
+  }, [dh, fetch]);
+
+  useHandleSceneTicks(model, containerRef.current);
+
+  return model ? (
+    <Chart
+      // eslint-disable-next-line react/jsx-props-no-spreading, @typescript-eslint/ban-ts-comment
+      // @ts-ignore
+      containerRef={containerRef}
+      model={model}
+      Plotly={Plotly}
+    />
+  ) : null;
+}
+
+export default PlotlyExpressChart;
