+++ plugins/plotly-express/src/js/src/index.ts
@@ -1,20 +1,6 @@
-import { type WidgetPlugin, PluginType } from '@deephaven/plugin';
-import { vsGraph } from '@deephaven/icons';
-import {
-  PlotlyExpressChart,
-  PlotlyExpressChartPanel,
-} from './PlotlyExpressChartPanel.js';
+import { PlotlyExpressPlugin } from './PlotlyExpressPlugin.js';
 
 export * from './PlotlyExpressChartModel.js';
 export * from './PlotlyExpressChartUtils.js';
 
-const plugin: WidgetPlugin = {
-  name: '@deephaven/plotly-express',
-  type: PluginType.WIDGET_PLUGIN,
-  supportedTypes: 'deephaven.plot.express.DeephavenFigure',
-  component: PlotlyExpressChart,
-  panelComponent: PlotlyExpressChartPanel,
-  icon: vsGraph,
-};
-
-export default plugin;
+export default PlotlyExpressPlugin;
